requirements.txt
# ============================================================================
# GEMINI Q&A SYSTEM - REQUIREMENTS (macOS Optimized)
# Python 3.11+ Compatible
# ============================================================================

# ----------------------------------------------------------------------------
# CORE PYTHON DEPENDENCIES
# ----------------------------------------------------------------------------
python-dotenv>=1.0.0
pydantic>=2.5.0,<3.0
pydantic-settings>=2.1.0
jsonschema>=4.21.0

# ----------------------------------------------------------------------------
# WEB FRAMEWORK & API
# ----------------------------------------------------------------------------
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.0.0

# ----------------------------------------------------------------------------
# DATABASE
# ----------------------------------------------------------------------------
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0,<3.0
alembic>=1.13.0

# ----------------------------------------------------------------------------
# AI/ML - PRIMARY MODELS
# ----------------------------------------------------------------------------
openai>=1.10.0
google-generativeai>=0.3.0
anthropic>=0.15.0

# ----------------------------------------------------------------------------
# AI/ML - TIER 2 LLM PROVIDERS
# ----------------------------------------------------------------------------
together>=1.2.0
groq>=0.4.0
replicate>=0.23.0
instructor>=0.4.0

# ----------------------------------------------------------------------------
# AI/ML - LOCAL MODELS (PyTorch CPU version for macOS)
# ----------------------------------------------------------------------------
--index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
--index-url https://pypi.org/simple
transformers>=4.36.0
sentence-transformers>=2.3.0

# ----------------------------------------------------------------------------
# LLM CONTROL
# ----------------------------------------------------------------------------
guidance>=0.1.10
outlines>=0.0.26

# ----------------------------------------------------------------------------
# AUDIO PROCESSING
# ----------------------------------------------------------------------------
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0
noisereduce>=3.0.0

# ----------------------------------------------------------------------------
# VIDEO/IMAGE PROCESSING
# ----------------------------------------------------------------------------
opencv-python>=4.9.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
pytesseract>=0.3.10
easyocr>=1.7.0
ultralytics>=8.1.0

# ----------------------------------------------------------------------------
# NLP & TEXT PROCESSING (pre-built wheels for macOS)
# ----------------------------------------------------------------------------
spacy>=3.7.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# ----------------------------------------------------------------------------
# DATA PROCESSING (REQUIRED)
# ----------------------------------------------------------------------------
pandas>=2.2.0
numpy>=1.26.0,<2.0.0  # Pin to avoid compatibility issues
openpyxl>=3.1.0
pyarrow>=14.0.0

# ----------------------------------------------------------------------------
# HTTP & NETWORKING
# ----------------------------------------------------------------------------
httpx>=0.26.0
requests>=2.31.0
aiohttp>=3.9.3
tenacity>=8.2.0
backoff>=2.2.0

# ----------------------------------------------------------------------------
# UTILITIES
# ----------------------------------------------------------------------------
tqdm>=4.66.0
python-magic>=0.4.27; sys_platform != 'darwin'  # Skip on macOS (requires libmagic)

# ----------------------------------------------------------------------------
# LOGGING & MONITORING
# ----------------------------------------------------------------------------
loguru>=0.7.0
python-json-logger>=2.0.0
rich>=13.7.0
colorlog>=6.8.0

# ----------------------------------------------------------------------------
# ADDITIONAL UTILITIES
# ----------------------------------------------------------------------------
python-dateutil>=2.8.0
pytz>=2024.1
click>=8.1.0
pyyaml>=6.0.0

# ----------------------------------------------------------------------------
# TESTING (Optional)
# ----------------------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# ============================================================================
# INSTALLATION NOTES
# ============================================================================
#
# macOS Installation:
#   pip install --index-url https://pypi.org/simple -r requirements.txt
#
# System Dependencies (install via Homebrew):
#   brew install ffmpeg
#   brew install tesseract  # Optional, for OCR
#
# Post-install:
#   python -m spacy download en_core_web_sm
#
# ============================================================================